feat: 15초 영상길이 옵션 + 한글 자막(스크립트 텍스트) 추가
- 영상 길이 선택지에 15초/30초 추가 (duration_minutes=0.25/0.5)
- text_draft_builder: 스크립트 텍스트를 한글 자막으로 하단 표시
  - 섹션 제목 → 상단(transform_y=0.7)
  - 핵심 포인트 → 중앙(transform_y=0.0)
  - 스크립트 자막 → 하단(transform_y=-0.8)
- 타이틀 카드 시간을 전체 길이의 10%로 adaptive 처리
- 총 시간 표시 초/분 단위 자동 전환

+++ core/script_template.py
@@ -23,7 +23,7 @@ _TEMPLATES = {
 }
 
 
-def generate_template(topic: str, duration_minutes: int = 5, num_sections: int = 3) -> Dict[str, Any]:
+def generate_template(topic: str, duration_minutes: float = 5, num_sections: int = 3) -> Dict[str, Any]:
     total_seconds = duration_minutes * 60
     templates = _TEMPLATES.get(num_sections, _TEMPLATES[3])
